Broke apart the Server monolith into a first draft with regards to code module organisation. I'm not entirely happy with this but it's a first step.
Tests.scala is a stab at showing how to develop client applications using this. There's a ton more to do in this area but I've sort of roughed out how to get at the Request and how to get at parameters.

True path and parameter matching has been added such that:
/foo/:bar/[:baz]
matches /foo/monkey/doo and /foo/monkey but not /foo or /foo/
That's bound to contain bugs but the general syntax is:
'/' [ {path-part} | {parameter-part} ] where
path-part is a (possibly empty) name with the usual valid identifiers
parameter-part is a name that starts with : (and optionally surrounded by [ and ] to denote optionality)