Some cleaning up of the Tests. Introduced a IntrinsicValues trait so that they (the intrinsics) can be offered-up as implicit values to enable fancy access to parameters. Right now you can:
val p = 'foo <=> "bar"

to indicate that p should be assigned the value of (optional) parameter foo (or "bar") if it was not supplied.

val q = 'foo <=

to indicate that q be assigned the value of parameter "foo". Will blow up if that parameter is optional and was not supplied.